+++ RELEASENOTES.md
@@ -1,1 +1,1 @@
-
+- Add description to dashboard